refactor(gosdr): add dataset module
Add GOES dataset metadata extraction module.
